Introduce `URIModelExportStore` abstraction
Define a new class `URIModelExportStore` that abstracts the common details of `BcoModelExportStore`, `ROCrateArchiveModelExportStore`, `TarModelExportStore` and `BagArchiveModelExportStore`.

This new class manages exports to file sources, where data can be retrieved later via a URI. It takes the responsibility of creating a temporary directory to set up the file to export and uploading it to the file source.